Move basic tensor rewriting code to aesara.tensor.rewriting
- `aesara.tensor.basic_opt` has been changed to `aesara.tensor.rewriting.basic`
- `aesara.tensor.math_opt` has been changed to `aesara.tensor.rewriting.math`
- `aesara.tensor.subtensor_opt` has been changed to `aesara.tensor.rewriting.subtensor`
- `aesara.tensor.opt_uncanonicalize` has been changed to `aesara.tensor.rewriting.uncanonicalize`

The tests associated with each module have been updated accordingly.
