+++ aesara/compile/builders.py
@@ -26,7 +26,7 @@ from aesara.graph.null_type import NullType
 from aesara.graph.op import HasInnerGraph, Op
 from aesara.graph.rewriting.basic import in2out, node_rewriter
 from aesara.graph.utils import MissingInputError
-from aesara.tensor.basic_opt import ShapeFeature
+from aesara.tensor.rewriting.basic import ShapeFeature
 
 
 def infer_shape(outs, inputs, input_shapes):
