+++ aesara/tensor/rewriting/__init__.py
@@ -1,0 +1,4 @@
+import aesara.tensor.rewriting.basic
+import aesara.tensor.rewriting.math
+import aesara.tensor.rewriting.subtensor
+import aesara.tensor.rewriting.uncanonicalize
