+++ aesara/tensor/rewriting/uncanonicalize.py
@@ -31,19 +31,14 @@ supposed to be canonical.
 
 """
 
-import logging
-
 from aesara import scalar as aes
 from aesara.graph.rewriting.basic import copy_stack_trace, node_rewriter
 from aesara.tensor.basic import Alloc, alloc, constant
-from aesara.tensor.basic_opt import register_uncanonicalize
 from aesara.tensor.elemwise import CAReduce, DimShuffle
 from aesara.tensor.math import Argmax, Max, MaxAndArgmax, Min, neg
+from aesara.tensor.rewriting.basic import register_uncanonicalize
 from aesara.tensor.shape import Reshape, reshape
 from aesara.tensor.subtensor import Subtensor
-
-
-_logger = logging.getLogger("aesara.tensor.opt_uncanonicalize")
 
 
 @register_uncanonicalize
