+++ aesara/tensor/utils.py
@@ -63,7 +63,7 @@ def shape_of_variables(fgraph, input_shapes):
     """
 
     if not hasattr(fgraph, "shape_feature"):
-        fgraph.attach_feature(aesara.tensor.basic_opt.ShapeFeature())
+        fgraph.attach_feature(aesara.tensor.rewriting.basic.ShapeFeature())
 
     input_dims = [
         dimension
