+++ setup.cfg
@@ -157,7 +157,7 @@ check_untyped_defs = False
 ignore_errors = True
 check_untyped_defs = False
 
-[mypy-aesara.tensor.basic_opt]
+[mypy-aesara.tensor.rewriting.basic]
 ignore_errors = True
 check_untyped_defs = False
 
@@ -193,7 +193,7 @@ check_untyped_defs = False
 ignore_errors = True
 check_untyped_defs = False
 
-[mypy-aesara.tensor.math_opt]
+[mypy-aesara.tensor.rewriting.math]
 ignore_errors = True
 check_untyped_defs = False
