+++ tests/tensor/random/test_basic.py
@@ -15,7 +15,6 @@ from aesara.graph.basic import Constant, Variable, graph_inputs
 from aesara.graph.fg import FunctionGraph
 from aesara.graph.op import get_test_value
 from aesara.graph.rewriting.db import RewriteDatabaseQuery
-from aesara.tensor.basic_opt import ShapeFeature
 from aesara.tensor.random.basic import (
     bernoulli,
     beta,
@@ -56,6 +55,7 @@ from aesara.tensor.random.basic import (
     wald,
     weibull,
 )
+from aesara.tensor.rewriting.basic import ShapeFeature
 from aesara.tensor.type import iscalar, scalar, tensor
 from tests.unittest_tools import create_aesara_param
