+++ tests/tensor/test_shape.py
@@ -11,8 +11,8 @@ from aesara.graph.type import Type
 from aesara.misc.safe_asarray import _asarray
 from aesara.tensor import as_tensor_variable, get_vector_length, row
 from aesara.tensor.basic import MakeVector, constant
-from aesara.tensor.basic_opt import ShapeFeature
 from aesara.tensor.elemwise import DimShuffle, Elemwise
+from aesara.tensor.rewriting.basic import ShapeFeature
 from aesara.tensor.shape import (
     Reshape,
     Shape_i,
